Avoid possible race condition in build_docs()
This addresses part of @hemkal's review comment in PR #13.

+++ leie/leie/serve.py
@@ -164,23 +164,29 @@ def requested_mimetype():
 
 def build_docs():
     """Build documentation in the current directory."""
-    with open("api.html.tmp", "w") as output:
-        try:
-            subprocess.run(["pandoc", "api.mdwn"], stdout=output)
-            os.rename("api.html.tmp", "api.html")
-        except FileNotFoundError as e:
-            os.remove("api.html.tmp")
-            s = ""
-            if os.path.exists("api.html"):
-                s = "re"
-            sys.stderr.write(
-                "WARNING: The 'pandoc' program was not available, so the\n")
-            sys.stderr.write(
-                "         'api.html' file has not been %sgenerated.\n" % s)
-            sys.stderr.write(
-                "         Install pandoc (https://pandoc.org/) to make\n")
-            sys.stderr.write(
-                "         this warning go away.\n")
+    # Doing this the old-fashioned way, instead of with 'with',
+    # because we'll be renaming the file.
+    output = open("api.html.tmp", "w")
+    try:
+        subprocess.run(["pandoc", "api.mdwn"], stdout=output)
+        output.close()
+        os.rename("api.html.tmp", "api.html")
+    except FileNotFoundError as e:
+        # 'pandoc' not installed, so warn and give up.
+        output.close() # this never ran in the 'try', so do it here
+        s = ""
+        if os.path.exists("api.html"):
+            s = "re"
+        sys.stderr.write( 
+            "WARNING: The 'pandoc' program was not available, so the\n")
+        sys.stderr.write(
+            "         'api.html' file has not been %sgenerated.\n" % s)
+        sys.stderr.write(
+            "         Install pandoc (https://pandoc.org/) to make\n")
+        sys.stderr.write(
+            "         this warning go away.\n")
+    if os.path.exists("api.html.tmp"):
+        os.remove("api.html.tmp")
 
 if __name__ == "__main__":
     # See https://github.com/SolutionGuidance/cavetl/issues/12 about
